Rename methods so they can't be confused with the @arguments decorator
