+++ mcmd/commands/add.py
@@ -4,9 +4,9 @@ from pathlib import Path
 
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.commands._registry import arguments
 from mcmd.client.molgenis_client import post, get, post_files
 from mcmd.command import command
+from mcmd.commands._registry import arguments
 from mcmd.io import highlight
 from mcmd.utils.errors import McmdError
 from mcmd.utils.file_helpers import get_file_name_from_path, scan_folders_for_files, select_path
@@ -20,7 +20,7 @@ p_add_theme = None
 # =========
 
 @arguments('add')
-def arguments(subparsers):
+def add_arguments(subparsers):
     global p_add_theme
     p_add = subparsers.add_parser('add',
                                   help='Add users and groups',
