+++ mcmd/commands/set.py
@@ -5,9 +5,9 @@ import json
 
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.commands._registry import arguments
 from mcmd.client.molgenis_client import get, put
 from mcmd.command import command
+from mcmd.commands._registry import arguments
 from mcmd.io import highlight
 from mcmd.utils.errors import McmdError
 
@@ -17,7 +17,7 @@ from mcmd.utils.errors import McmdError
 # =========
 
 @arguments('set')
-def arguments(subparsers):
+def add_arguments(subparsers):
     p_set = subparsers.add_parser('set',
                                   help='Alter settings',
                                   description="Run 'mcmd set -h' to view the help for this sub-command")
